Upgrade to react-diagrams v6.2
I'm not aware of anything we actually need from this new release, but
we've only just started using react-diagrams, so we might as well use
an up-to-date version.

As part of this I added a bunch of peer dependencies that Yarn was
complaining about, because I don't think there's a need for them to be
exposed as peer dependencies from this project.