Fix installation of node dependencies in CI
I'd managed to misunderstand how the CircleCI config works, and wasn't
properly installing the modules before testing. Somehow this had been
passing in CI up until now because there was a cached tree available
and the install wasn't necessary because the necessary state was
cached, but that stopped working now.