Fix background colour and font size for AST diagrams
The font size is set to an explicit value because the layout will
depend on the calculated size of the node boxes, and if the font size
is allowed to change based on the styles of the page this is embedded
in, then the boxes will not be spaced out properly.

Setting the background colour for the whole diagram means that it will
always render OK even when embedding in places that have a different
background colour.