Set the background colour only on the nodes
I decided that we could get away with only setting the background
colour on the AST nodes. If the diagram is embedded in somewhere with
an off-white background colour it will still look correct.

This won't work so  well with embedding somewhere with a dark
background colour, because the links between nodes won't be
visible. But in that case the whole style sheet can be overridden.